requirements.txt
boto3>=1.26.0
botocore>=1.29.0
python-dateutil>=2.8.2
python-dotenv>=1.0.0
Pillow>=9.0.0
pytest>=7.0.0
pytest-cov>=4.0.0
moto>=4.0.0
requests>=2.28.0
awscli-local>=0.20
